requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
easyocr>=1.7.0
numpy>=1.24.0
Pillow>=9.0.0
websockets>=12.0
aiofiles>=23.0.0
scipy>=1.11.0
deep-sort-realtime>=1.3.2
torch>=2.0.0
torchvision>=0.15.0
albumentations==1.3.1
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
